it's alive!
I wanted to make the Engine a separate thing in case I want to put
a different frontend on it. The python client is just a testing thing at
this point, I would wrap this with a better API so you could do things
like `client.set("foo", "bar")` and the wrapper would do the needful
underneath, but it works! it's a start!