make exported methods not interface with the underlying storage directly
one of my next tasks is going to be expiration, so sticking to a harder
interface around the storage will allow me to do fun things like
automatically expire things upon fetching, keeping a list of keys that
have expiration, etc.

I have a feeling I'll even extract the storage itself completely out of
the engine at some point, but we'll see. Right now it's a matter of
discipline to not muck around directly with the storage, but pulling it
out to its own type would make it so I have to stick to the exported
interface.