refactor NewEngine just a bit
refactor NewEngine to use an internal method newEngine that returns an
engine so I can test the engine directly and make new non-exported
methods on it, but still export the Engine interface.

This also tests situations where unsupported data types are used in
certain methods to make sure they behave as expected. The fakeValueStore
struct is that unsupported data type :)