change lots of return values
added an IntValue protobuf message and converted the commands that
return ints in redis to return ints here as well. also made Type() able
to return an error even though it ... can't? error? just in case, for
future compatibility. Every engine method should return an error.

I'm pleased with this.