tests for the server and some other things
this has:
* unit tests for the server
* which uncovered an issue with my thinking around SingleValue
responses, given that they have a default value if unset, and I need to
be able to differentiate between set to the default value and explicitly
unset, so this is the suggestion I found online.

* an unimplemented engine "base" type. So not all engines have to
implement everything. Hooray.

* my first foray into mocking. I also like how I'm laying out this
server, there's the server, then the engine, and you specify the engine
to the server (making for pluggable engines if you wanted) and whatnot.
I have tried to keep the internals of the engine separatable, too, so as
to maybe have a "storage" interface inside the engine (or at least this
engine) that can be swapped in and out as well. Who knows.

It's not a full suite of tests for the server, but it's getting there.