first attempt at integration tests
I'm abandoning this approach for the moment because I'm flailing and
don't really know what I should be doing. I thought I'd extract out the
actual run-the-server-application to its own package and put the
integration tests there, but I also need some helpers for the engine for
fake values and the like and thought I'd extract those to their own
package, but the problem is the engine itself is not exported (only the
Engine protocol and the NewEngine method) so I can't monkeypatch it in
tests in a different package and I've gone too far down this rabbithole
and gotten lost and I want to back out and take another stab at it.

saving this work in a branch for posterity, though, as usual