factor out a stringishValueStoreInterface
I decided I didn't want to have a super bloated valueStore interface and
a "base class" that implemented error responses for unsupported methods
and that I wanted different interfaces for the different valueStore
types

originally having it all on one was fine because I distilled down what
all of the stringish functions I'd implemented to 2 functions on the
valueStore (get and incrby) but with sets there are just so many more
functions and such and I think I'd rather break it apart. This pushes
type checking into the main engine API, which isn't great, but might be
able to be DRYed out a different way. We shall see if I end up thinking
this was a good idea or not.

Somewhat related, but I'm really starting to feel like I should break up
string, int, and float value stores into separate stores and sets of
functions. Thinking about the shenanigans required to juggle switching
between types and ... anyways.

Also one fun thing this does get us is I should be able to reuse
stringishValueStoreInterface structs in the hash value store type later
if I want :D